requirements.txt
# requirements.txt for Autoclass AI Agent

fastapi
uvicorn[standard]
requests
sqlalchemy
psycopg2-binary
python-dotenv